+++ backend/src/main/java/codesquad/team01/issuetracker/issue/repository/IssueQueryRepository.java
@@ -8,7 +8,7 @@ import codesquad.team01.issuetracker.issue.dto.IssueDto;
 public interface IssueQueryRepository {
 
 	List<IssueDto.BaseRow> findIssuesWithFilters(
-		IssueState state, Long writerId, Long milestoneId,
-		List<Long> labelIds, List<Long> assigneeIds);
+		IssueState state, Integer writerId, Integer milestoneId,
+		List<Integer> labelIds, List<Integer> assigneeIds);
 
 }
