+++ backend/src/main/java/codesquad/team01/issuetracker/issue/service/IssueAssembler.java
@@ -23,11 +23,11 @@ public class IssueAssembler {
 		log.debug("이슈 {}개에 대한 상세 정보 조합", issues.size());
 
 		// 담당자 정보를 이슈 id로 그룹화
-		Map<Long, List<UserDto.AssigneeResponse>> assigneesByIssueId = groupAssigneesByIssueId(assignees);
+		Map<Integer, List<UserDto.AssigneeResponse>> assigneesByIssueId = groupAssigneesByIssueId(assignees);
 		log.debug("이슈별 담당자 정보 그룹화: {}개 이슈-담당자 매핑", assigneesByIssueId.size());
 
 		// 레이블 정보를 이슈 id로 그룹화
-		Map<Long, List<LabelDto.ListItemResponse>> labelsByIssueId = groupLabelsByIssueId(labels);
+		Map<Integer, List<LabelDto.ListItemResponse>> labelsByIssueId = groupLabelsByIssueId(labels);
 		log.debug("이슈별 레이블 정보 그룹화: {}개 이슈-레이블 매핑", labelsByIssueId.size());
 
 		// 각 이슈 정보에 담당자, 레이블 결합
@@ -41,7 +41,7 @@ public class IssueAssembler {
 	}
 
 	// 담당자 정보 이슈 id로 그룹화
-	private Map<Long, List<UserDto.AssigneeResponse>> groupAssigneesByIssueId(
+	private Map<Integer, List<UserDto.AssigneeResponse>> groupAssigneesByIssueId(
 		List<UserDto.IssueAssigneeRow> assignees) {
 		return assignees.stream()
 			.collect(Collectors.groupingBy(
@@ -57,7 +57,7 @@ public class IssueAssembler {
 	}
 
 	// 레이블 정보 이슈 id로 그룹화
-	private Map<Long, List<LabelDto.ListItemResponse>> groupLabelsByIssueId(List<LabelDto.IssueLabelRow> labels) {
+	private Map<Integer, List<LabelDto.ListItemResponse>> groupLabelsByIssueId(List<LabelDto.IssueLabelRow> labels) {
 		return labels.stream()
 			.collect(Collectors.groupingBy(
 				LabelDto.IssueLabelRow::issueId,
